requirements.txt
python-dotenv
fireworks-ai
pydantic